all update(qna 제외)

+++ src/main/java/service/review/ReviewDeleteService.java
@@ -9,8 +9,10 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.apache.ibatis.session.SqlSession;
 
+import common.C;
 import domain.ReviewWriteDAO;
 import domain.ReviewWriteDTO;
+import domain.UserDTO;
 import service.Service;
 import sqlmapper.SqlSessionManager;
 
@@ -27,7 +29,16 @@ public class ReviewDeleteService implements Service {
 		
 		try {
 			sqlSession = SqlSessionManager.getInstance().openSession();
-			dao = sqlSession.getMapper(ReviewWriteDAO.class);
+			dao = sqlSession.getMapper(ReviewWriteDAO.class);			
+			
+			// 로그인 한 사용자가 아니면 여기서 redirect 해야 한다
+			UserDTO logUser = (UserDTO)request.getSession().getAttribute(C.PRINCIPAL);
+			List<ReviewWriteDTO> list = dao.selectById(id);
+			UserDTO writeUser = list.get(0).getUser();
+			if(logUser.getId() != writeUser.getId()) {
+				response.sendRedirect(request.getContextPath() + "/user/user_rejectAuth");
+				return;
+			}
 			
 			cnt = dao.deleteById(id);
